# doc/OpenAPI2/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "API used to manage the extended documents in UDV Server.",
    "version": "1.0.0",
    "title": "Spatial-Multimedia-DB API$"
  },
  "host": "localhost:1525",
...
  "schemes": [
    "https"
  ],
  "paths": {
    "/login": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Log in and retrieve a token",
        "description": "Authenticate a user by giving them a auth token",
        "consumes": [
          "multipart/form-data"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "username",
            "in": "formData",
            "description": "Username of the user",
            "required": true,
            "type": "string"
          },
          {
            "name": "password",
            "in": "formData",
            "description": "Password of the user",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully logged in",
            "schema": {
              "$ref": "#/definitions/TokenItem"
            }
          },
          "401": {
            "description": "Authentication is needed to perform the request"
          },
          "500": {
            "description": "Unexpected server error (should not happen)"
          }
        }
      }
    },
    "/user": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Create a new user",
        "description": "Add a new User by passing informations below\n",
        "consumes": [
          "multipart/form-data"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "lastName",
            "in": "formData",
            "description": "Pass the lastName of user you want to create",
            "required": true,
            "type": "string"
          },
          {
            "name": "firstName",
            "in": "formData",
            "description": "Pass the firstName of user you want to create",
            "required": true,
            "type": "string"
          },
          {
            "name": "email",
            "in": "formData",
            "description": "Pass the email of user you want to create",
            "required": true,
            "type": "string"
          },
          {
            "name": "username",
            "in": "formData",
            "description": "Pass the username of user you want to create",
            "required": true,
            "type": "string"
          },
          {
            "name": "password",
            "in": "formData",
            "description": "Pass the password of user you want to create",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "201": {
            "description": "Information about the new user",
            "schema": {
              "$ref": "#/definitions/UserItem"
            }
          },
          "422": {
            "description": "Unprocessable entity (invalid field, duplicate item)\n"
          },
          "500": {
            "description": "Unexpected server error (should not happen)"
          }
        }
      }
    },
    "/user/me": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Retrieve the connected user",
        "description": "Retrieve the logged in user with the given token\n",
        "produces": [
          "application/json"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "The connected user's information",
            "schema": {
              "$ref": "#/definitions/UserItem"
            }
          },
          "401": {
            "description": "Authentication is needed to perform the request"
          },
          "500": {
            "description": "Unexpected server error (should not happen)"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/user/{id}": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Retrieve user by their ID",
        "description": "Retrieve a specific User by passing the id\n",
        "consumes": [
          "multipart/form-data"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Pass the id of user you want to retrieve",
            "required": true,
            "type": "integer",
            "x-example": 1
          }
        ],
        "responses": {
          "200": {
            "description": "The user's information",
            "schema": {
              "$ref": "#/definitions/UserItem"
            }
          },
          "404": {
            "description": "The specified resource was not found"
          },
          "500": {
            "description": "Unexpected server error (should not happen)"
          }
        }
      }
    },
    "/user/grant": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Create a privileged user",
        "description": "Create a user with the specific role",
        "consumes": [
          "multipart/form-data"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "lastName",
            "in": "formData",
            "description": "Pass the lastName of user you want to create",
            "required": true,
            "type": "string"
          },
          {
            "name": "firstName",
            "in": "formData",
            "description": "Pass the firstName of user you want to create",
            "required": true,
            "type": "string"
          },
          {
            "name": "email",
            "in": "formData",
            "description": "Pass the email of user you want to create",
            "required": true,
            "type": "string"
          },
          {
            "name": "username",
            "in": "formData",
            "description": "Pass the username of user you want to create",
            "required": true,
            "type": "string"
          },
          {
            "name": "password",
            "in": "formData",
            "description": "Pass the password of user you want to create",
            "required": true,
            "type": "string"
          },
          {
            "name": "role",
            "in": "formData",
            "description": "The role of the user",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "The user's information",
            "schema": {
              "$ref": "#/definitions/UserItem"
            }
          },
          "404": {
            "description": "The specified resource was not found"
          },
          "500": {
            "description": "Unexpected server error (should not happen)"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/document": {
      "get": {
        "tags": [
          "Documents"
        ],
        "summary": "Get all extended documents",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successfully get the extended documents",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/ExtendedDocument"
              }
            }
          },
          "500": {
            "description": "Unexpected server error (should not happen)"
          }
        }
      },
      "post": {
        "tags": [
          "Documents"
        ],
        "summary": "Create a new extended document",
        "consumes": [
          "multipart/form-data"
        ],
        "parameters": [
          {
            "name": "title",
            "in": "formData",
            "required": true,
            "type": "string"
          },
          {
            "name": "description",
            "in": "formData",
            "required": true,
            "type": "string"
          },
          {
            "name": "refDate",
            "in": "formData",
            "required": true,
            "type": "string"
          },
          {
            "name": "publicationDate",
            "in": "formData",
            "required": true,
            "type": "string"
          },
          {
            "name": "source",
            "in": "formData",
            "required": true,
            "type": "string"
          },
          {
            "name": "rightsHolder",
            "in": "formData",
            "required": false,
            "type": "string"
          },
          {
            "name": "quaternionX",
            "in": "formData",
            "required": true,
            "type": "number"
          },
          {
            "name": "quaternionY",
            "in": "formData",
            "required": true,
            "type": "number"
          },
          {
            "name": "quaternionZ",
            "in": "formData",
            "required": true,
            "type": "number"
          },
          {
            "name": "quaternionW",
            "in": "formData",
            "required": true,
            "type": "number"
          },
          {
            "name": "positionX",
            "in": "formData",
            "required": true,
            "type": "number"
          },
          {
            "name": "positionY",
            "in": "formData",
            "required": true,
            "type": "number"
          },
          {
            "name": "positionZ",
            "in": "formData",
            "required": true,
            "type": "number"
          },
          {
            "name": "file",
            "in": "formData",
            "required": true,
            "type": "file"
          }
        ],
        "responses": {
          "201": {
            "description": "Successfully created an extended document",
            "schema": {
              "$ref": "#/definitions/ExtendedDocument"
            }
          },
          "400": {
            "description": "Request is malformed (a mandatory field is missing)"
          },
          "401": {
            "description": "Authentication is needed to perform the request"
          },
          "403": {
            "description": "The user does not has sufficient privilege to perform the request\n"
          },
          "415": {
            "description": "The file has an incorrect format\n"
          },
          "422": {
            "description": "Unprocessable entity (invalid field, duplicate item)\n"
          },
          "500": {
            "description": "Unexpected server error (should not happen)"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/document/{id}": {
      "get": {
        "tags": [
          "Documents"
        ],
        "summary": "Get an extended document with a specific id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the extendedDoc.",
            "required": true,
            "type": "integer",
            "x-example": 1
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully get the extendedDoc",
            "schema": {
              "$ref": "#/definitions/ExtendedDocument"
            }
          },
          "404": {
            "description": "The specified resource was not found"
          },
          "500": {
            "description": "Unexpected server error (should not happen)"
          }
        }
      },
      "put": {
        "tags": [
          "Documents"
        ],
        "summary": "Update an extended document",
        "consumes": [
          "multipart/form-data"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the document.",
            "required": true,
            "type": "integer",
            "x-example": 1
          },
          {
            "name": "title",
            "in": "formData",
            "required": false,
            "type": "string"
          },
          {
            "name": "description",
            "in": "formData",
            "required": false,
            "type": "string"
          },
          {
            "name": "refDate",
            "in": "formData",
            "required": false,
            "type": "string"
          },
          {
            "name": "publicationDate",
            "in": "formData",
            "required": false,
            "type": "string"
          },
          {
            "name": "source",
            "in": "formData",
            "required": false,
            "type": "string"
          },
          {
            "name": "rightsHolder",
            "in": "formData",
            "required": false,
            "type": "string"
          },
          {
            "name": "quaternionX",
            "in": "formData",
            "required": false,
            "type": "string"
          },
          {
            "name": "quaternionY",
            "in": "formData",
            "required": false,
            "type": "string"
          },
          {
            "name": "quaternionZ",
            "in": "formData",
            "required": false,
            "type": "string"
          },
          {
            "name": "quaternionW",
            "in": "formData",
            "required": false,
            "type": "string"
          },
          {
            "name": "positionX",
            "in": "formData",
            "required": false,
            "type": "string"
          },
          {
            "name": "positionY",
            "in": "formData",
            "required": false,
            "type": "string"
          },
          {
            "name": "positionZ",
            "in": "formData",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully updated the document",
            "schema": {
              "$ref": "#/definitions/ExtendedDocument"
            }
          },
          "401": {
            "description": "Authentication is needed to perform the request"
          },
          "403": {
            "description": "The user does not has sufficient privilege to perform the request\n"
          },
          "404": {
            "description": "The specified resource was not found"
          },
          "422": {
            "description": "Unprocessable entity (invalid field, duplicate item)\n"
          },
          "500": {
            "description": "Unexpected server error (should not happen)"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Documents"
        ],
        "summary": "Delete an extended document with a specific id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the document.",
            "required": true,
            "type": "integer",
            "x-example": 1
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully deleted the document",
            "schema": {
              "$ref": "#/definitions/ExtendedDocument"
            }
          },
          "401": {
            "description": "Authentication is needed to perform the request"
          },
          "403": {
            "description": "The user does not has sufficient privilege to perform the request\n"
          },
          "404": {
            "description": "The specified resource was not found"
          },
          "500": {
            "description": "Unexpected server error (should not happen)"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/document/{id}/file": {
      "get": {
        "tags": [
          "Document files"
        ],
        "summary": "Get the file associated to a document",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the extendedDoc.",
            "required": true,
            "type": "integer",
            "x-example": 1
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully get the document's image with the specified id\n"
          },
          "404": {
            "description": "The specified resource was not found"
          },
          "500": {
            "description": "Unexpected server error (should not happen)"
          }
        }
      },
      "post": {
        "tags": [
          "Document files"
        ],
        "summary": "Upload and associate a file to document",
        "consumes": [
          "multipart/form-data"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the document",
            "required": true,
            "type": "integer",
            "x-example": 1
          },
          {
            "name": "file",
            "in": "formData",
            "description": "The file to upload",
            "required": true,
            "type": "file"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully uploaded the document's file"
          },
          "401": {
            "description": "Authentication is needed to perform the request"
          },
          "403": {
            "description": "The user does not has sufficient privilege to perform the request\n"
          },
          "404": {
            "description": "The specified resource was not found"
          },
          "422": {
            "description": "Unprocessable entity (invalid field, duplicate item)\n"
          },
          "500": {
            "description": "Unexpected server error (should not happen)"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Document files"
        ],
        "summary": "Delete the file associated to a document",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the extendedDoc.",
            "required": true,
            "type": "integer",
            "x-example": 1
          }
        ],
        "responses": {
          "204": {
            "description": "Successfully delete the extendedDocument file"
          },
          "401": {
            "description": "Authentication is needed to perform the request"
          },
          "403": {
            "description": "The user does not has sufficient privilege to perform the request\n"
          },
          "404": {
            "description": "The specified resource was not found"
          },
          "500": {
            "description": "Unexpected server error (should not happen)"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/document/in_validation": {
      "get": {
        "tags": [
          "Document validation"
        ],
        "summary": "Get the documents in validation",
        "parameters": [],
        "responses": {
          "200": {
            "description": "The retrieved documents",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/ExtendedDocument"
              }
            }
          },
          "401": {
            "description": "Authentication is needed to perform the request"
          },
          "403": {
            "description": "The user does not has sufficient privilege to perform the request\n"
          },
          "500": {
            "description": "Unexpected server error (should not happen)"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
...
    "/link": {
      "get": {
        "tags": [
          "Links"
        ],
        "summary": "Retrieve the available target types for a link",
        "consumes": [
          "multipart/form-data"
        ],
        "responses": {
          "200": {
            "description": "Successfully get the possible target types",
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "example": "city_object"
              }
            }
          },
          "500": {
            "description": "Unexpected server error (should not happen)"
          }
        }
      }
    },
...
  },
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "definitions": {
    "TokenItem": {
      "type": "object",
      "properties": {
        "token": {
          "type": "string",
          "format": "jwt",
          "example": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJ1c2VyX2lkIjoxLCJ1c2VybmFtZSI6ImFkbWluIiwiZmlyc3ROYW1lIjoiQWRtaW4iLCJsYXN0TmFtZSI6IlVEViIsImVtYWlsIjoiYWRtaW5AdWR2LmZyIiwicG9zaXRpb24iOnsiTEVWRUxfTUlOIjoxLCJsYWJlbCI6ImFkbWluIiwiaWQiOjEsImNsZWFyYW5jZSI6WyJjb250cmlidXRvciIsInNvZnRNb2RlcmF0b3IiLCJtb2RlcmF0b3IiLCJhZG1pbiJdLCJMRVZFTF9NQVgiOjN9LCJleHAiOjE1NTg3NzcwMzYuMDM1MTczfQ.-6ns6_K-VpRlsqTrXbrV4shm8Oa0GJn3IfiR0LmLJoE"
        }
      }
    },
...
    "UserItem": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "example": 1
        },
        "username": {
          "type": "string",
          "example": "jdoe"
        },
        "firstName": {
          "type": "string",
          "example": "John"
        },
        "lastName": {
          "type": "string",
          "example": "Doe"
        },
        "email": {
          "type": "string",
          "example": "john.doe@example.com"
        },
        "position": {
          "$ref": "#/definitions/PositionItem"
        },
        "position_id": {
          "type": "integer",
          "example": 4
        }
      }
    },
...
  },
...
}